Añadido al controlador de órdenes la opcion de expand en listas y el control de que exista el registro para las operaciones de get, update y delete

+++ src/controllers/order.controller.js
@@ -19,6 +19,29 @@ export default class OrderController extends BaseController {
             query += this.createOrderByCaluse (req, Order.visibleFields());
             query += this.createLimitClause(req);
             const result = await this.query(query, queryParameters);
+            if (req.query.expand && req.query.expand.toLowerCase() === "true") {
+                const orderLinesByOrder = {};
+                if(result && result.length > 0){
+                    const orderIds = [];
+                    for(const orderResult of result){
+                        const order = new Order(orderResult);
+                        orderIds.push(order.id);
+                    }
+                    const orderLinesResult = await this.query("SELECT " + OrderLine.visibleFields().join(', ') + " FROM " + OrderLine.table() + " WHERE requestId IN (?)", orderIds);
+                    if(orderLinesResult && orderLinesResult.length > 0){
+                        for(const orderLineResult of orderLinesResult){
+                            const orderLine = new OrderLine(result);
+                            if(!orderLinesByOrder[orderLine.requestId]){
+                                orderLinesByOrder[orderLine.requestId] = []
+                            }
+                            orderLinesByOrder[orderLine.requestId].push(orderLineResult);
+                        }
+                    }
+                    for(const orderResult of result){
+                        orderResult.lines = orderLinesByOrder[orderResult.id] || [];
+                    }
+                }
+            }
             return response.success('Orders Retrieved successfully', result);
         } catch (error) {
             console.log(error);
@@ -31,12 +54,13 @@ export default class OrderController extends BaseController {
         try {
             const { id } = req.params;
             const result = await this.query("SELECT " + Order.visibleFields().join(', ') + " FROM " + Order.table() + " WHERE id = ?", id);
-            const order = new Order(result);
-            if (req.query.expand) {
-                if (req.query.expand.toLowerCase() === "true") {
-                    const orderLines = await this.query("SELECT " + OrderLine.visibleFields().join(', ') + " FROM " + OrderLine.table() + " WHERE requestId = ?", id);
-                    order.lines = Utils.forceArray(orderLines);
-                }
+            const order = result && result.length === 1 ? new Order(result) : undefined;
+            if(!order){
+                return response.notFound('Not found order with id ' + id);
+            } 
+            if (req.query.expand && req.query.expand.toLowerCase() === "true") {
+                const orderLines = await this.query("SELECT " + OrderLine.visibleFields().join(', ') + " FROM " + OrderLine.table() + " WHERE requestId = ?", id);
+                order.lines = Utils.forceArray(orderLines);
             }
             return response.success('Order Retrieved successfully', order);
         } catch (error) {
@@ -48,7 +72,6 @@ export default class OrderController extends BaseController {
         const response = new HTMLResponse(req, res);
         try {
             const order = new Order(req.body);
-
             if (order.date == undefined) {
                 return response.badRequest('Missing one of these fields: date', HTMLResponse.MISSING_DATA_STATUS)
             }
@@ -70,7 +93,11 @@ export default class OrderController extends BaseController {
         try {
             const { id } = req.params;
             const order = new Order(req.body);
-    
+            const searchResult = await this.query("SELECT " + Order.visibleFields().join(', ') + " FROM " + Order.table() + " WHERE id = ?", id);
+            const searchOrder = result && result.length === 1 ? new Order(searchResult) : undefined;
+            if(!searchOrder){
+                return response.notFound('Not found order with id ' + id);
+            } 
             if (order.date == undefined) {
                 return response.badRequest('Missing one of these fields: date', HTMLResponse.MISSING_DATA_STATUS)
             }
@@ -90,6 +117,11 @@ export default class OrderController extends BaseController {
         const response = new HTMLResponse(req, res);
         try {
             const { id } = req.params;
+            const searchResult = await this.query("SELECT " + Order.visibleFields().join(', ') + " FROM " + Order.table() + " WHERE id = ?", id);
+            const searchOrder = result && result.length === 1 ? new Order(searchResult) : undefined;
+            if(!searchOrder){
+                return response.notFound('Not found order with id ' + id);
+            } 
             const result = await this.query("DELETE FROM " + Order.table() + " WHERE id = ?", id);
             return response.success('Order deleted successfully');
         } catch (error) {
